refactor(transform): split oxc processor application
