+++ packages/transform/src/utils/applyOxcProcessors/displayName.ts
@@ -1,0 +1,167 @@
+/* eslint-disable no-restricted-syntax */
+
+import { basename, dirname } from 'path';
+
+import type { ExpressionValue } from '@wyw-in-js/shared';
+import { ValueType } from '@wyw-in-js/shared';
+import type { Node, Program } from 'oxc-parser';
+
+import type { OxcStaticValue } from '../collectOxcTemplateDependencies';
+import { isOxcNode, walkOxc } from '../oxc/ast';
+import { isNodeReference } from './cleanupBindings';
+import type { AnyNode } from './types';
+
+export const getPropertyKeyName = (
+  property: AnyNode,
+  code: string
+): string | null => {
+  const { key } = property;
+  if (!isOxcNode(key)) {
+    return null;
+  }
+
+  if (key.type === 'Identifier') {
+    return key.name;
+  }
+
+  if (key.type === 'Literal') {
+    return String(key.value);
+  }
+
+  return typeof key.start === 'number' && typeof key.end === 'number'
+    ? code.slice(key.start, key.end)
+    : null;
+};
+
+export const getDisplayName = (
+  ancestors: Node[],
+  idx: number,
+  code: string,
+  filename?: string | null
+): string => {
+  const owner = [...ancestors].reverse().find((node) => {
+    return (
+      node.type === 'Property' ||
+      node.type === 'JSXOpeningElement' ||
+      node.type === 'VariableDeclarator'
+    );
+  }) as AnyNode | undefined;
+
+  if (owner?.type === 'Property') {
+    const keyName = getPropertyKeyName(owner, code);
+    if (keyName) {
+      return keyName;
+    }
+  } else if (owner?.type === 'JSXOpeningElement') {
+    const { name } = owner;
+    if (isOxcNode(name) && name.type === 'JSXIdentifier') {
+      return name.name;
+    }
+  } else if (owner?.type === 'VariableDeclarator') {
+    const { id } = owner;
+    if (isOxcNode(id) && id.type === 'Identifier') {
+      return id.name;
+    }
+  }
+
+  let displayName = basename(filename ?? 'unknown').replace(/\.[a-z\d]+$/, '');
+  if (filename && /^index\.[a-z\d]+$/.test(basename(filename))) {
+    displayName = basename(dirname(filename));
+  }
+
+  if (!displayName) {
+    throw new Error(
+      "Couldn't determine a name for the component. Ensure that it's either:\n" +
+        '- Assigned to a variable\n' +
+        '- Is an object property\n' +
+        '- Is a prop in a JSX element\n'
+    );
+  }
+
+  return `${displayName}${idx}`;
+};
+
+export const getTagOwner = (ancestors: Node[]): AnyNode | null => {
+  const owner = [...ancestors]
+    .reverse()
+    .find(
+      (node) =>
+        node.type === 'Property' ||
+        node.type === 'JSXOpeningElement' ||
+        node.type === 'VariableDeclarator'
+    ) as AnyNode | undefined;
+
+  return owner ?? null;
+};
+
+export const isTagReferenced = (
+  program: Program,
+  ancestors: Node[]
+): boolean => {
+  const owner = getTagOwner(ancestors);
+  if (owner?.type !== 'VariableDeclarator') {
+    return true;
+  }
+
+  const { id } = owner;
+  if (!isOxcNode(id) || id.type !== 'Identifier') {
+    return true;
+  }
+
+  if (ancestors.some((node) => node.type === 'ExportNamedDeclaration')) {
+    return true;
+  }
+
+  let referenced = false;
+  walkOxc(program, (node, parent) => {
+    const referenceName =
+      node.type === 'Identifier' || node.type === 'JSXIdentifier'
+        ? node.name
+        : null;
+
+    if (
+      referenced ||
+      referenceName !== id.name ||
+      (node.type === 'Identifier' &&
+        node.start === id.start &&
+        node.end === id.end)
+    ) {
+      return;
+    }
+
+    referenced = isNodeReference(node, parent);
+  });
+
+  return referenced;
+};
+
+export const collectSameFileProcessorStaticValues = (
+  expressionValues: Omit<ExpressionValue, 'buildCodeFrameError'>[],
+  processorStaticValuesByLocal: Map<string, unknown>
+): OxcStaticValue[] => {
+  const staticValues: OxcStaticValue[] = [];
+  const seen = new Set<string>();
+
+  expressionValues.forEach((value) => {
+    if (value.kind !== ValueType.LAZY) {
+      return;
+    }
+
+    const staticValue = processorStaticValuesByLocal.get(value.source);
+    if (
+      staticValue === undefined ||
+      value.ex.type !== 'Identifier' ||
+      seen.has(value.ex.name)
+    ) {
+      return;
+    }
+
+    seen.add(value.ex.name);
+    staticValues.push({
+      name: value.ex.name,
+      value: staticValue,
+    });
+  });
+
+  return staticValues;
+};
