+++ scripts/check-production-ts-size.mjs
@@ -16,7 +16,6 @@ const LEGACY_OVERSIZED_FILES = new Map([
     'packages/transform/src/transform/generators/resolveStaticOxcValues.ts',
     6084,
   ],
-  ['packages/transform/src/utils/applyOxcProcessors.ts', 2937],
   ['packages/transform/src/utils/collectOxcExportsAndImports.ts', 1436],
   ['packages/transform/src/utils/oxcPreevalTransforms.ts', 1794],
   ['packages/transform/src/utils/oxcShaker.ts', 1082],
